refactor(archive): move legacy files to archive
- Move files used during FDD/TDD phases to archive
- Consolidate functions into domain layer during DDD refactoring
  e.g., moved validation logic into domain models
- Keep code clean by excluding unused files
