+++ src/main/kotlin/lotto/archive/CalcProfitRate.kt
@@ -1,4 +1,4 @@
-package lotto.service.logic
+package lotto.archive
 
 fun calcProfitRate(totalWinningAmount: Int, totalLottoPurchaseAmount: Int): Double {
     return totalWinningAmount.toDouble() / totalLottoPurchaseAmount.toDouble() * 100
