+++ src/main/kotlin/lotto/archive/ValidateWinningMain.kt
@@ -1,4 +1,4 @@
-package lotto.service.validation
+package lotto.archive
 
 /**
  *  [ Background ]
