+++ src/test/kotlin/lotto/archive/CalcProfitRateTest.kt
@@ -1,11 +1,11 @@
-package lotto.service.logic
+package lotto.archive
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.Assertions.assertEquals
 
 class CalcProfitRateTest {
     @Test
-    fun `givenZeroWinningAmount_whenCalculateProfitRate_thenReturns0%`() {
+    fun `givenZeroWinningAmount_whenCalculateProfitRate_thenReturns0Percent`() {
         val totalWinningAmount = 0
         val totalLottoPurchaseAmount = 10000
         val expected = 0.0
@@ -14,7 +14,7 @@ class CalcProfitRateTest {
     }
 
     @Test
-    fun `givenTotalWinningAmountAndTotalLottoPurchaseAmount_whenCalcProfitRate_thenReturns50%`() {
+    fun `givenTotalWinningAmountAndTotalLottoPurchaseAmount_whenCalcProfitRate_thenReturns50Percent`() {
         val totalWinningAmount = 5000
         val totalLottoPurchaseAmount = 10000
         val expected = 50.0
@@ -23,7 +23,7 @@ class CalcProfitRateTest {
     }
 
     @Test
-    fun `givenWinningAmountEqualToPurchaseAmount_whenCalcProfitRate_thenReturns100%`() {
+    fun `givenWinningAmountEqualToPurchaseAmount_whenCalcProfitRate_thenReturns100Percent`() {
         val totalWinningAmount = 10000
         val totalLottoPurchaseAmount = 10000
         val expected = 100.0
@@ -32,7 +32,7 @@ class CalcProfitRateTest {
     }
 
     @Test
-    fun `givenWinningAmountGreaterThanPurchaseAmount_whenCalcProfitRate_thenReturns200%`() {
+    fun `givenWinningAmountGreaterThanPurchaseAmount_whenCalcProfitRate_thenReturns200Percent`() {
         val totalWinningAmount = 20000
         val totalLottoPurchaseAmount = 10000
         val expected = 200.0
