+++ src/test/kotlin/lotto/archive/RoundProfitRateTest.kt
@@ -1,4 +1,4 @@
-package lotto.util
+package lotto.archive
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.Assertions.assertEquals
