+++ src/test/kotlin/lotto/archive/ValidateLottoNumbersTest.kt
@@ -1,4 +1,4 @@
-package lotto.service.validation
+package lotto.archive
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
